# src/InterfaceAdapters/CleanArchitecture.Api/Requests/UpdateSampleRequestModel.cs
namespace CleanArchitecture.Api.Requests;

/// <summary>
/// Request Model
/// </summary>
#pragma warning disable CS1591
public class UpdateSampleRequestModel
{
    public int Id { get; set; }
    public required string Name { get; set; }
}